Move orderbook specific log function out of logger.ts
